+++ blocks/header/menu/hashTagParser.js
@@ -1,3 +1,5 @@
+import { isDesktop } from './menu.js';
+
 /**
  * Parses href hash
  *
@@ -24,7 +26,7 @@ function parseHashTag(href) {
   const hashTags = href.split('#').slice(1);
 
   if (hashTags) {
-    // @TODO: optimize (now it is going ot be O(n^2) :/)
+    // @TODO: optimize (now it is going to be O(n^2) :/)
     namespaces.forEach((ns) => {
       hashTags.forEach((tag) => {
         const value = tag.split(ns);
@@ -40,6 +42,46 @@ function parseHashTag(href) {
   return parsed;
 }
 
-export {
-  parseHashTag,
-};
+
+/**
+ * Applies parsed hash tags based on conditions
+ *
+ * @param aElement
+ * @param hashTags
+ */
+function applyConditions(aElement, hashTags) {
+  hashTags.forEach((hashTag) => {
+    const { namespace, value } = { ...hashTag };
+
+    if (namespace === 'display_for_') {
+      if (value === 'desktop_only' && !isDesktop.matches) {
+        aElement.parentNode.removeChild(aElement);
+      }
+      if (value === 'mobile_only' && isDesktop.matches) {
+        aElement.parentNode.removeChild(aElement);
+      }
+    }
+  });
+}
+
+/**
+ * Parses nav fragment; disable visible elements based on hash tags
+ *
+ * @param navSections
+ * @returns {*}
+ */
+function parseUrlHashTags(navSections) {
+  const aElements = navSections.querySelectorAll('li > a');
+
+  aElements.forEach((aElement) => {
+    const link = aElement.href;
+    if (link) {
+      const hashTags = parseHashTag(link);
+      console.table(`Hash tags for ${link}`, hashTags);
+
+      applyConditions(aElement, hashTags);
+    }
+  });
+}
+
+export default parseUrlHashTags;
